rename files from test.js
