+++ packages/kit/test/build-errors/apps/server-only-folder-dynamic-import/src/lib/server/something/explode.js
