+++ packages/kit/test/build-errors/apps/server-only-folder-dynamic-import/src/routes/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-	const mod = import('$lib/server/something/test.js');
+	const mod = import('$lib/server/something/explode.js');
 </script>
 
 {#await mod then resolved}
