+++ packages/kit/test/build-errors/apps/server-only-folder/src/lib/server/something/explode.js
