+++ packages/kit/test/build-errors/apps/server-only-folder/src/routes/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-	import { should_explode } from '$lib/server/something/test.js';
+	import { should_explode } from '$lib/server/something/explode.js';
 </script>
 
 <p>{should_explode}</p>
